resolver: merge HashBuilder-s and ResolveSet-s, decouple from ResolveRequest-s (#540)
* merge hbStorage and hb in resolver

* Fix prepare

* Remove tracing by default

* cleanup

* Fix linter

* Not to include incarnation into the ResolveRequest

* Fix linter

* Fix linter

* Fix test

* Fix linter

* Revert

* Fix for infinite loop

* Fix linter

* Collapse rss into rs

* remove reliance of requests in resolver

* Remove currentReq

* Remove fields from ResolveRequest

* Fix linter

+++ core/state/dump.go
@@ -197,7 +197,7 @@ func (d *Dumper) dump(c collector, excludeCode, excludeStorage, _ bool, start []
 				dbutils.CurrentStateBucket,
 				dbutils.StorageHistoryBucket,
 				storagePrefix,
-				8*uint(common.HashLength+common.IncarnationLength),
+				8*(common.HashLength+common.IncarnationLength),
 				d.blockNumber,
 				func(ks, vs []byte) (bool, error) {
 					storageMap[common.BytesToHash(ks[common.HashLength:])] = common.CopyBytes(vs)
